+++ CHANGES.rst
@@ -8,6 +8,7 @@ in progress
 - Feature: Copy test data to InfluxDB
 - Tests: Speed up test data import by specifying row count
 - Tests: Add test for ``influxio info``
+- Feature: Add reading line protocol format from file
 
 
 2023-xx-xx 0.1.0